+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/GenerateGraphQLSchemaMojo.java
@@ -35,7 +35,7 @@ import graphql.ThreadSafe;
  * 
  * @author etienne-sf
  */
-@Mojo(name = "generateGraphQLSchema", defaultPhase = LifecyclePhase.INITIALIZE, requiresProject = true)
+@Mojo(name = "generateGraphQLSchema", defaultPhase = LifecyclePhase.INITIALIZE, requiresProject = true, threadSafe = true)
 @ThreadSafe
 // No need to add the @Component spring annotation: AbstractCommonMojo added this instance into the spring context, to
 // use the instance which attributs has been set with the pom content