+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/GeneratePojoMojo.java
@@ -59,7 +59,7 @@ import graphql.ThreadSafe;
  * 
  * @author etienne-sf
  */
-@Mojo(name = "generatePojo", defaultPhase = LifecyclePhase.GENERATE_SOURCES, requiresProject = true)
+@Mojo(name = "generatePojo", defaultPhase = LifecyclePhase.GENERATE_SOURCES, requiresProject = true, threadSafe = true)
 @ThreadSafe
 // No need to add the @Component spring annotation: AbstractCommonMojo added this instance into the spring context, to
 // use the instance which attributes has been set with the pom content